Fix problem where target assignment sends not all events.
